TM4/requirements.txt
openai>=1.30.0
chromadb>=0.5.0
python-dotenv>=1.0.1